Refactoring component and using prebuilt log
